+++ crates/wdl-engine/CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 #### Added
 
+* Added support for `disks` requirement mount points in the Docker backend
+  ([#527](https://github.com/stjude-rust-labs/sprocket/pull/528)). The mount
+  points are created as volumes but size constraints cannot be enforced.
 * Added setting for controlling content digests; supported values are `strong`
   for full cryptographic hashing of file content and `weak` to digest based
   solely off file metadata ([#503](https://github.com/stjude-rust-labs/sprocket/pull/503)).